% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Script_PLATE_04_PREPROCESS_3_1_CheckAlignment_PSI.R
\name{CheckAlignment.PSI}
\alias{CheckAlignment.PSI}
\title{Check splicing data}
\usage{
CheckAlignment.PSI(MarvelObject)
}
\arguments{
\item{MarvelObject}{Marvel object. S3 object generated from \code{CreateMarvelObject} function.}
}
\value{
An object of class S3 with updated slots \code{MarvelObject$SplicePheno}, \code{MarvelObject$SpliceFeature}, and \code{MarvelObject$PSI}.
}
\description{
Checks if the metadata aligns with the columns and rows of the matrix for splicing data.
}
\examples{
marvel.demo <- readRDS(system.file("extdata/data", "marvel.demo.rds", package="MARVEL"))

marvel.demo <- CheckAlignment.PSI(MarvelObject=marvel.demo)
}
