% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Script_PLATE_04_PREPROCESS_3_4_CheckAlignment_SJ.R
\name{CheckAlignment.SJ}
\alias{CheckAlignment.SJ}
\title{Check splice junction data}
\usage{
CheckAlignment.SJ(MarvelObject)
}
\arguments{
\item{MarvelObject}{Marvel object. S3 object generated from \code{CreateMarvelObject} function.}
}
\value{
An object of class S3 with updated slots \code{MarvelObject$SplicePheno}, \code{MarvelObject$PSI} and \code{MarvelObject$IntronCounts}.
}
\description{
Checks if the metadata aligns with the columns and rows of the matrix for splice junction data prior to PSI computation.
}
\examples{
marvel.demo <- readRDS(system.file("extdata/data", "marvel.demo.rds", package="MARVEL"))

marvel.demo <- CheckAlignment.SJ(MarvelObject=marvel.demo)
}
