% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Script_PLATE_08_DE_1_3_CompareValues_Gene.R
\name{CompareValues.Exp}
\alias{CompareValues.Exp}
\title{Differential gene expression analysis}
\usage{
CompareValues.Exp(
  MarvelObject,
  cell.group.g1,
  cell.group.g2,
  downsample = FALSE,
  min.cells = 25,
  pct.cells = NULL,
  method,
  method.adjust,
  show.progress = TRUE,
  nboots = 1000
)
}
\arguments{
\item{MarvelObject}{Marvel object. S3 object generated from \code{TransformExpValues} function.}

\item{cell.group.g1}{Vector of character strings. Cell IDs corresponding to Group 1 (reference group).}

\item{cell.group.g2}{Vector of character strings. Cell IDs corresponding to Group 2.}

\item{downsample}{Logical value. If set to \code{TRUE}, the number of cells in each cell group will be downsampled to the sample size of the smaller cell group so that both cell groups will have the sample size prior to differential expression analysis. Default is \code{FALSE}.}

\item{min.cells}{Numeric value. The minimum no. of cells expressing the gene for the gene to be included for differential splicing analysis.}

\item{pct.cells}{Numeric value. The minimum no. of cells expressing the gene for the gene to be included for differential splicing analysis. If \code{pct.cells} is specified, then \code{pct.cells} will be used as threshold instead of \code{min.cells}.}

\item{method}{Character string. Statistical test to compare the 2 groups of cells. \code{"ks"}, \code{"kuiper"}, \code{"ad"}, \code{"dts"}, \code{"wilcox"}, and \code{"t.test"} for Kolmogorov-Smirnov, Kuiper, Anderson-Darling, DTS, Wilcox, and t-test, respectively.}

\item{method.adjust}{Character string. Adjust p-values for multiple testing. Options available as per \code{p.adjust} function.}

\item{show.progress}{Logical value. If set to \code{TRUE}, progress bar will be displayed so that users can estimate the time needed for differential analysis. Default value is \code{TRUE}.}

\item{nboots}{Numeric value. When \code{method} set to \code{"dts"}, the number of bootstrap iterations for computing the p-value.}
}
\value{
An object of class S3 new slot \code{MarvelObject$DE$Exp$Table}.
}
\description{
Performs differential gene expression analysis between 2 groups of cells.
}
