% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Script_DROPLET_04_DE_2_Gene.R
\name{CompareValues.Genes.10x}
\alias{CompareValues.Genes.10x}
\title{Differential gene expression analysis}
\usage{
CompareValues.Genes.10x(
  MarvelObject,
  log2.transform = TRUE,
  show.progress = TRUE
)
}
\arguments{
\item{MarvelObject}{Marvel object. S3 object generated from \code{CompareValues.SJ.10x} function.}

\item{log2.transform}{Logical value. If set to \code{TRUE} (default), normalised gene expression values will be off-set by 1 and then log2-transformed prior to analysis.}

\item{show.progress}{Logical value. If set to \code{TRUE} (default), the progress bar will appear.}
}
\value{
An object of class S3 with a updated slot \code{MarvelObject$DE$SJ$Table}.
}
\description{
Performs differential gene expression analysis between two groups of cells. Only among cells and genes previously included for splice junction analysis.
}
