% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Script_DROPLET_06_DE_PSEUDOBULK_1_SJ.R
\name{CompareValues.SJ.DonorLevel.10x}
\alias{CompareValues.SJ.DonorLevel.10x}
\title{Differential splice junction analysis at donor level}
\usage{
CompareValues.SJ.DonorLevel.10x(
  MarvelObject,
  cell.group.list,
  min.pct.cells.gene.expr = 10,
  min.n.cells.gene.expr = 10,
  min.gene.counts.total = 3,
  min.sj.count = 1,
  min.donor.gene.expr = 3,
  min.donor.sj.expr = 3
)
}
\arguments{
\item{MarvelObject}{cell.group.list List of character strings. List defining the two cell groups and their respective donor IDs and their corresponding cells IDs.}

\item{min.pct.cells.gene.expr}{Numeric value. Mininum percentage of cells across a donor expressing a given gene for that gene to be considered expressed. Genes below this threshold will be considered not expressed. Default is \code{10}. To be used in conjunction with \code{min.n.cells.gene.expr} and \code{min.gene.counts.total}.}

\item{min.n.cells.gene.expr}{Numeric value. Mininum number of cells across a donor expressing a given gene for that gene to be considered expressed. Genes below this threshold will be considered not expressed. Default is \code{10}. To be used in conjunction with \code{min.pct.cells.gene.expr} and \code{min.gene.counts.total}.}

\item{min.gene.counts.total}{Numeric value. Mininum number of UMI counts across a donor for that gene to be considered expressed. Genes below this threshold will be considered not expressed. Default is \code{3}. To be used in conjunction with \code{min.n.cells.gene.expr} and \code{min.pct.cells.gene.expr}.}

\item{min.sj.count}{Numeric value. Mininum number of UMI counts across a donor for that splice junction to be considered expressed. Splice junctions below this threshold will be considered not expressed. Default is \code{1}.}

\item{min.donor.gene.expr}{Numeric value. Minimum number of donors expressing a given gene based on \code{min.pct.cells.gene.expr}, \code{min.n.cells.gene.expr}, and \code{min.gene.counts.total} options for that gene to be included for analysis. Default is \code{3}.}

\item{min.donor.sj.expr}{Numeric value. Minimum number of donors expressing a given splice junction based on \code{min.sj.count} option for that splice junction to be included for analysis. Default is \code{3}.}
}
\value{
An object of class S3 with a new slots \code{MarvelObject$DE$Pseudobulk$SJ$Table} and \code{MarvelObject$DE$Pseudobulk$SJ$cell.group.list}.
}
\description{
Performs differential splice junction analysis between two groups of cells by aggregating cells by their corresponding donor IDs. Statistical test used here is the permutation test.
}
