% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Script_PLATE_04_COMPUTE_PSI_8_Compute_Posterior.R
\name{ComputePSI.Posterior}
\alias{ComputePSI.Posterior}
\title{Compute PSI with Bayesian approach}
\usage{
ComputePSI.Posterior(MarvelObject)
}
\arguments{
\item{MarvelObject}{S3 object generated from \code{ComputePSI} function.}
}
\value{
An object of class S3 with new slots \code{$Counts}.
}
\description{
Combines a prior with likelihood (observed SJ counts) to obtain a posterior value of PSI.
}
\details{
This function computes posterior value of PSI based on a prior and likelihood (observed SJ counts). The resulting PSI values may be subsequently used for downstream principal component analysis (PCA). Because a prior is used here, there are no arguments to set minimum coverage threshold for computing the PSI values.
}
