% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Script_DROPLET_01_CREATE_MARVEL_OBJECT.R
\name{CreateMarvelObject.10x}
\alias{CreateMarvelObject.10x}
\title{Create Marvel object for droplet-based RNA-sequencing data}
\usage{
CreateMarvelObject.10x(
  gene.norm.matrix = NULL,
  gene.norm.pheno = NULL,
  gene.norm.feature = NULL,
  gene.count.matrix = NULL,
  gene.count.pheno = NULL,
  gene.count.feature = NULL,
  sj.count.matrix = NULL,
  sj.count.pheno = NULL,
  sj.count.feature = NULL,
  pca = NULL,
  gtf = NULL
)
}
\arguments{
\item{gene.norm.matrix}{Sparse matrix. UMI-collapsed, normalised, non-log2-transformed gene expression matrix.}

\item{gene.norm.pheno}{Data frame. Sample metadata for annotating \code{gene.norm.matrix} columns with cell IDs.}

\item{gene.norm.feature}{Data frame. Gene metadata for annotating \code{gene.norm.matrix} rows with gene names.}

\item{gene.count.matrix}{Sparse matrix. UMI-collapsed, non-normalised (raw counts), non-log2-transformed gene expression matrix.}

\item{gene.count.pheno}{Data frame. Sample metadata for annotating \code{gene.count.matrix} columsn with cell IDs.}

\item{gene.count.feature}{Data frame. Gene metadata for annotating \code{gene.count.matrix} rows with gene names.}

\item{sj.count.matrix}{Sparse matrix. UMI-collapsed, non-normalised (raw counts), non-log2-transformed splice junction expression matrix.}

\item{sj.count.pheno}{Data frame. Sample metadata for annotating \code{sj.count.matrix} columsn with cell IDs.}

\item{sj.count.feature}{Data frame. Splice junction metadata for annotating \code{sj.count.matrix} rows with splice junction coordinates.}

\item{pca}{Data frame. Coordinates of PCA/tSNE/UMAP.}

\item{gtf}{Data frame. GTF used in cellranger. Will be used for annotating splice junctions downstream.}
}
\value{
An object of class S3.
}
\description{
Creates an S3 object named \code{Marvel} for downstream analysis, specifically for droplet-based RNA-sequencing data.
}
