% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Script_DROPLET_06_DE_Pseudobulk_1_SJ_CreatePseudobulk.R
\name{CreatePseudoBulk.SJ.10x}
\alias{CreatePseudoBulk.SJ.10x}
\title{Create pseudobulk for downstream differential splicing analysis}
\usage{
CreatePseudoBulk.SJ.10x(
  MarvelObject,
  donor.id.column,
  feature.column,
  comparison.column,
  min.pct.cells.genes = 10
)
}
\arguments{
\item{MarvelObject}{Marvel object. S3 object generated from \code{CheckAlignment.10x} function.}

\item{donor.id.column}{Character strings. Column name of the sample metadata the corresponds to the donor ID for pseudo-bulking.}

\item{feature.column}{Character strings. Column name of the sample metadata the corresponds to the cell type for pseudo-bulking.}

\item{comparison.column}{Character strings. Column name of the sample metadata the corresponds to the cell groups for downstream pair-wise comparison.}

\item{min.pct.cells.genes}{Numeric value. Minimum percentage of cells in which the gene is expressed for that gene to be included for pseudobulk-ing. Expressed genes defined as genes with non-zero normalised UMI counts.}
}
\value{
An object of class S3 with a new slots \code{MarvelObject$DE$PseudoBulk$GeneCountMatrix}, \code{MarvelObject$DE$PseudoBulk$SJCountMatrix}, and \code{MarvelObject$DE$PseudoBulk$sample.metadata}.
}
\description{
Create pseudobulk by merging gene counts and SJ counts by donor for each cell type. The results in each cell type from each donor having aggregated gene counts and SJ counts.
}
