% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Script_PLATE_08_DE_1_1_CompareValues_DownsamplebyGenes.R
\name{DownsampleByGenes}
\alias{DownsampleByGenes}
\title{Downsample cells based on number of expressed genes}
\usage{
DownsampleByGenes(
  MarvelObject,
  cell.group.g1,
  cell.group.g2,
  sliding.window = 50
)
}
\arguments{
\item{MarvelObject}{Marvel object. S3 object generated from \code{TransformExpValues} function.}

\item{cell.group.g1}{Vector of character strings. Cell IDs corresponding to Group 1 (reference group).}

\item{cell.group.g2}{Vector of character strings. Cell IDs corresponding to Group 2.}

\item{sliding.window}{Numeric values. The matching cell to look for with the number of genes detected within plus minus of this specified value.}
}
\value{
An object of class S3 new slots \code{MarvelObject$DE$Exp$Downsampled.Data$Table}, \code{MarvelObject$DE$Exp$Downsampled.Data$Plot.Before.Downsample}, and \code{MarvelObject$DE$Exp$Downsampled.Data$Plot.After.Downsample}.
}
\description{
Match the cells from both cell groups based on the number of expressed genes.
}
