% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Script_PLATE_08_PCA_3_ElbowPlot.R
\name{ElbowPlot}
\alias{ElbowPlot}
\title{Elbow plot}
\usage{
ElbowPlot(MarvelObject, level, n.dim = 50)
}
\arguments{
\item{MarvelObject}{Marvel object. S3 object generated from \code{RunPCA} function.}

\item{level}{Character strings. Indicate \code{"splicing"} or \code{"gene"} to retrieve the eigenvalues from splicing or gene expression principal component analysis (PCA), respectively.}

\item{n.dim.}{Indicate the first number of dimensions to plot the eigenvalues for. Default value is \code{50}, i.e., the first 50 PCs.}
}
\value{
An object of class S3 containing with new slots \code{ MarvelObject$PCA$PSI$ElbowPlot} or \code{MarvelObject$PCA$Exp$ElbowPlot} when \code{level} option set to \code{"splicing"} or \code{"gene"}, respectively.
}
\description{
Create elbw plot from principal components (PCs).
}
