% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comics.R
\name{get_comics}
\alias{get_comics}
\title{Get information on Marvel comics.}
\usage{
get_comics(id = NULL, offset = NULL, limit = NULL)
}
\arguments{
\item{id}{unique comics id. Optional.}

\item{offset}{The requested number of skipped results of the call. Optional.}

\item{limit}{The requested results limit. Optional.}
}
\value{
a dataframe containing the comics' information
}
\description{
The get_comics API returns information on Marvel comics
Optionally, users can specify the ID of a comics.
Users can also specify the offset and limit of the response.
If there is only one returned character, get_comics API will return a dataframe containing the
detailed information of the comics. If there are several returned comics, this API will only
return brief information on each comics.
}
\examples{
get_comics()
get_comics(id = 59548)
get_comics(offset = 50, limit = 3)
}
