% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_creators.R
\name{get_creators}
\alias{get_creators}
\title{Get information on Marvel creators.}
\usage{
get_creators(id = NULL, offset = NULL, limit = NULL)
}
\arguments{
\item{id}{unique creator id. Optional.}

\item{offset}{The requested number of skipped results of the call. Optional.}

\item{limit}{The requested results limit. Optional.}
}
\value{
a dataframe containing the creator(s)' information
}
\description{
The get_creators API returns information on Marvel creators.
Optionally, users can specify the ID of a creator.
Users can also specify the offset and limit of the response.
If there is only one returned creator, get_comics API will return a dataframe containing the
detailed information of the creator If there are several returned creators, this API will only
return brief information on each comics.
}
\examples{
get_creators()
get_creators(id = 3188)
get_creators(offset = 50, limit = 3)
}
