% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_events.R
\name{get_events}
\alias{get_events}
\title{Get information on Marvel events.}
\usage{
get_events(id = NULL, name = NULL, offset = NULL, limit = NULL)
}
\arguments{
\item{id}{unique creator id. Optional.}

\item{name}{title of an event. Optional.}

\item{offset}{The requested number of skipped results of the call. Optional.}

\item{limit}{The requested results limit. Optional.}
}
\value{
a dataframe containing the event(s)' information
}
\description{
The get_events API returns information on Marvel events.
Optionally, users can specify the ID of an event or the name of an event.
Users can also specify the offset and limit of the response.
If there is only one returned event, get_comics API will return a dataframe containing the
detailed information of the event. If there are several returned events, this API will only
return brief information on each event.
}
\examples{
get_events()
get_events(id = 321)
get_events(name = "Spider-Verse")
get_events(offset = 50, limit = 3)
}
