% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_series.R
\name{get_series}
\alias{get_series}
\title{Get information on Marvel series.}
\usage{
get_series(id = NULL, offset = NULL, limit = NULL)
}
\arguments{
\item{id}{unique creator id. Optional.}

\item{offset}{The requested number of skipped results of the call. Optional.}

\item{limit}{The requested results limit. Optional.}
}
\value{
a dataframe containing the event(s)' information
}
\description{
The get_series API gets information on Marvel series.
Optionally, users can specify the ID of an event.
Users can also specify the offset and limit of the response.
If there is only one returned event, get_series API will return a dataframe containing the
detailed information of the event. If there are several returned series, this API will only
return brief information on each series.
}
\examples{
get_series()
get_series(id = 1234)
get_series(offset = 50, limit = 3)
}
