% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stories.R
\name{get_stories}
\alias{get_stories}
\title{Get information on Marvel stories.}
\usage{
get_stories(id = NULL, offset = NULL, limit = NULL)
}
\arguments{
\item{id}{unique creator id. Optional.}

\item{offset}{The requested number of skipped results of the call. Optional.}

\item{limit}{The requested results limit. Optional.}
}
\value{
a dataframe containing the event(s)' information
}
\description{
The get_stories API gets information on Marvel stories
Optionally, users can specify the ID of a story.
Users can also specify the offset and limit of the response.
If there is only one returned story, get_stories API will return a dataframe containing the
detailed information of the story. If there are several returned stories, this API will only
return brief information on each story.
}
\examples{
get_stories()
get_stories(id = 12)
get_stories(offset = 1000, limit = 3)
}
