% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Script_PLATE_08_PCA_5_ComputeHighlyVariable_PSI.R
\name{IdentifyVariableEvents}
\alias{IdentifyVariableEvents}
\title{Identify highly variable splicing events}
\usage{
IdentifyVariableEvents(
  MarvelObject,
  sample.ids = NULL,
  cell.group.column,
  cell.group.order,
  min.cells = 25
)
}
\arguments{
\item{MarvelObject}{Marvel object. S3 object generated from \code{ComputePSI} function.}

\item{sample.ids}{Character strings. Specific cells to plot.}

\item{cell.group.column}{Character string. The name of the sample metadata column in which the variables will be used included for analysis}

\item{cell.group.order}{Character string. The the variables under the sample metadata column specified in \code{cell.group.column} to be included for analysis}

\item{min.cells}{Numeric value. The minimum no. of cells expressing the splicing event, above which, the splicing event will be retained for analysis.}
}
\value{
An object of class S3 containing with new slots \code{ MarvelObject$VariableSplicing$tran_ids} and  \code{MarvelObject$VariableSplicing$plot}
}
\description{
Identify highly variable splicing events by fitting spline model. The vector splicing events returned may be used for downstream dimension reduction analysis with the \code{RunPCA} function.
}
