% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Script_PLATE_08_DE_2_3_PlotDEValues_PSI_Mean_Group1vs2.R
\name{PlotDEValues.PSI.Mean.g2vsg1}
\alias{PlotDEValues.PSI.Mean.g2vsg1}
\title{Plot differential splicing analysis results based on mean PSI difference}
\usage{
PlotDEValues.PSI.Mean.g2vsg1(
  MarvelObject,
  method,
  pval,
  delta = 5,
  point.size = 1,
  xlabel.size = 8,
  anno = FALSE,
  anno.tran_id = NULL,
  label.size = 2.5,
  point.alpha = 1,
  event.types = c("SE", "MXE", "RI", "A5SS", "A3SS", "AFE", "ALE"),
  event.types.colors = NULL
)
}
\arguments{
\item{MarvelObject}{Marvel object. S3 object generated from \code{CompareValues} function.}

\item{method}{Character string. The statistical method used for differential splicing analysis.}

\item{pval}{Numeric value. Adjusted p-value below which the splcing event are considered as statistically significant and will consequently be color-annotated on the plot.}

\item{delta}{Numeric value. The positive (and negative) value specified above (and below) which the splicing events are considered to be statistically significant and will consequently be color-annotated on the plot.}

\item{point.size}{Numeric value. The point size for the data points. Default value is \code{1}.}

\item{xlabel.size}{Numeric value. Font size of the xtick labels. Default is \code{8}.}

\item{anno}{Logical value. If set to \code{TRUE}, the specific gene names will be annotated on the plot. Speficified together with \code{anno.tran_id}.}

\item{anno.tran_id}{Vector of character strings. When \code{anno} set to \code{TRUE}, the coordinates of the splicing events to be annotated on the plot.}

\item{label.size}{Numeric value. Only applicable if \code{anno} set to TRUE. Size of the gene name labels.}

\item{point.alpha}{Numeric value. Transpancy of data points. Default is \code{1}.}

\item{event.types}{Vector of character string(s). The specific splicing event to plot. May take any one or more of the following values \code{"SE"}, \code{"MXE"}, \code{"RI"}, \code{"A5SS"}, \code{"A3SS"}, \code{"AFE"}, and  \code{"ALE"}.}

\item{event.types.colors}{Vector of character string(s). Customise colors as per splicing event type specified in \code{event.types} option. Should be of same length as \code{event.types} option.}
}
\value{
An object of class S3 containing with new slot \code{MarvelObject$DE$PSI$Plot[["method"]]}.
}
\description{
Scatterplot of differential splicing analysis results based on mean PSI difference between 2 groups of cells. x-axis represents the mean PSI values of cell group 1. y-axis represents the mean PSI values of cell group 2.
}
\examples{
marvel.demo <- readRDS(system.file("extdata/data", "marvel.demo.rds", package="MARVEL"))

marvel.demo <- PlotDEValues.PSI.Mean.g2vsg1(MarvelObject=marvel.demo,
                                     method="ad",
                                     pval=0.10,
                                     delta=5
                                     )

# Check output
marvel.demo$DE$PSI$Plot
marvel.demo$DE$PSI$Summary
}
