% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Script_PLATE_10_ADHOC_PLOT_2_PlotValues_Gene.R
\name{PlotValues.Exp}
\alias{PlotValues.Exp}
\title{Plot gene expression values}
\usage{
PlotValues.Exp(
  MarvelObject,
  cell.group.list,
  feature,
  maintitle = "gene_short_name",
  xlabels.size = 8,
  cell.group.colors = NULL
)
}
\arguments{
\item{MarvelObject}{Marvel object. S3 object generated from \code{TransformExpValues} function.}

\item{cell.group.list}{List of character strings. Each element of the list is a vector of cell IDs corresponding to a cell group. The name of the element will be the cell group label.}

\item{feature}{Character string. \code{gene_id} for plotting. Should match \code{gene_id} column of \code{MarvelObject$GeneFeature} slot.}

\item{maintitle}{Character string. Column to use as plot main title as per \code{MarvelObject$GeneFeature}. Default is \code{"gene_short_name"} column.}

\item{xlabels.size}{Numeric value. Size of x-axis labels as per \code{ggplot2} function. Default is 8.}

\item{cell.group.colors}{Character string. Vector of colors for the cell groups specified for PCA analysis using \code{cell.type.columns}, \code{cell.type.variable}, and \code{cell.type.labels}. If not specified, default \code{ggplot2} colors will be used.}
}
\value{
An object of class S3 with new slot \code{MarvelObject$adhocPlot$Exp}.
}
\description{
Boxplot of gene expression values across different groups of cells.
}
