% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Script_DROPLET_07_ADHOC_PLOT_BOXPLOT_2_PlotValues_Gene_SingleCell.R
\name{PlotValues.Gene.SingleCell.10x}
\alias{PlotValues.Gene.SingleCell.10x}
\title{Violin plot of single-cell PSI values vs cell groups}
\usage{
PlotValues.Gene.SingleCell.10x(
  MarvelObject,
  cell.group.list,
  gene_short_name,
  cell.group.colors = NULL,
  xlabels.size = 8,
  alpha = 0.5
)
}
\arguments{
\item{MarvelObject}{Marvel object. S3 object generated from \code{CheckAlignment.10x} function.}

\item{cell.group.list}{List of character strings. Each element in the list is a cell group. The label of each element of the list represents the cell group label.}

\item{gene_short_name}{Character string. Gene name whose expression will be plotted.}

\item{cell.group.colors}{Vector of character strings. Colors of cell groups and should be same length as \code{cell.group.list}. Default \code{ggplot2} colors are used.}

\item{xlabels.size}{Numeric value. Font size of x-tick labels. Default is \code{10}.}
}
\value{
An object of class S3 with new slots \code{MarvelObject$adhocPlot$ViolinPlot$SingleCell$Plot} and \code{MarvelObject$adhocPlot$ViolinPlot$SingleCell$Data}.
}
\description{
Violin plot of single-cell PSI values on the y-axis against cell groups on the x-axis.
}
