% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Script_DROPLET_06_ADHOC_PLOT_PCA_3_PlotValues_Gene.R
\name{PlotValues.PCA.Gene.10x}
\alias{PlotValues.PCA.Gene.10x}
\title{Annotate reduced dimension space with gene expression values}
\usage{
PlotValues.PCA.Gene.10x(
  MarvelObject,
  gene_short_name,
  point.size = 0.1,
  color.gradient = c("grey90", "blue", "red"),
  type
)
}
\arguments{
\item{MarvelObject}{Marvel object. S3 object generated from \code{CheckAlignment.10x} function.}

\item{gene_short_name}{Character string. Gene name whose expression will be plotting.}

\item{point.size}{Numeric value. Size of data points. Default is \code{1}.}

\item{color.gradient}{Vector of character strings. Colors to indicate low, moderate, and high expression. Default is \code{c("grey90","blue","red")}.}

\item{type}{Character string. Type of reduced dimension space. Options are \code{"umap"} and \code{"tsne"}.}
}
\value{
An object of class S3 with new slot \code{MarvelObject$adhocPlot$PCA$Gene}.
}
\description{
Annotates reduced dimension space, e.g., UMAP and tSNE, with gene expression values. Values will be automatically be log2-transformed prior to plotting.
}
