% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Script_DROPLET_08_ADHOC_PLOT_BOXPLOT_1_PlotValues_PseudoBulk_PSI.R
\name{PlotValues.PSI.Pseudobulk.10x}
\alias{PlotValues.PSI.Pseudobulk.10x}
\title{Boxplot of pseudo-bulk PSI values vs cell groups}
\usage{
PlotValues.PSI.Pseudobulk.10x(
  MarvelObject,
  cell.group.list,
  coord.intron,
  cell.group.colors = NULL,
  xlabels.size = 10,
  min.pct.cells.gene.expr = 10,
  min.n.cells.gene.expr = 3,
  min.gene.counts.total = 3,
  method = "t.test",
  p.adjust.method = "fdr"
)
}
\arguments{
\item{MarvelObject}{Marvel object. S3 object generated from \code{CheckAlignment.10x} function.}

\item{cell.group.list}{List of character strings. Each element in the list (main list) is a list (sub-list). Each sub-list in turn is a list of pseudobulk samples represented by cell IDs. Each element in the main list is a cell group and the label of each element of the main list represents the cell group label.}

\item{coord.intron}{Character string. Coordinates of splice junction whose expression will be plotted.}

\item{cell.group.colors}{Vector of character strings. Colors of cell groups and should be same length as \code{cell.group.list}. Default \code{ggplot2} colors are used.}

\item{xlabels.size}{Numeric value. Font size of x-tick labels. Default is \code{10}.}

\item{min.pct.cells.gene.expr}{Numeric value. Percentage of cell expressing the gene in a pseudobulk, below which, the pseudobulk will be omitted from plotting. Default is \code{10}.}

\item{min.n.cells.gene.expr}{Numeric value. Number of cell expressing the gene in a pseudobulk, below which, the pseudobulk will be omitted from plotting. Default is \code{3}.}

\item{min.gene.counts.total}{Numeric value. Totol gene counts in a pseudobulk, below which, the pseudobulk will be omitted from plotting. Default is \code{3}.}

\item{method}{Character string. Statistical test for all possible pair-wise comparisons. Options are \code{"t.test"} (default) or \code{"wilcox"}.}

\item{p.adjust.method}{Character string. Method for multiple testing adjustment as per \code{method} option of \code{p.adjust} function. Default is \code{"fdr"}.}
}
\value{
An object of class S3 with new slots \code{MarvelObject$adhocPlot$Boxplot$Pseudobulk$PSI$Plot}, \code{MarvelObject$adhocPlot$Boxplot$Pseudobulk$PSI$Stats}, and \code{MarvelObject$adhocPlot$Boxplot$Pseudobulk$PSI$Data}.
}
\description{
Boxplot of pseudo-bulk PSI values on the y-axis against cell groups on the x-axis.
}
