% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Script_DROPLET_08_ADHOC_PLOT_HEATMAP_PlotValues_PSI_PseudoBulk.R
\name{PlotValues.PSI.Pseudobulk.Heatmap.10x}
\alias{PlotValues.PSI.Pseudobulk.Heatmap.10x}
\title{Heatmap of pseudo-bulk PSI values across different cell groups}
\usage{
PlotValues.PSI.Pseudobulk.Heatmap.10x(
  MarvelObject,
  coord.intron,
  x.column,
  y.column,
  x.levels,
  y.levels,
  xlabels.size = 8,
  ylabels.size = 8,
  min.pct.cells.gene.expr = 10,
  min.n.cells.gene.expr = 10,
  min.gene.counts.total = 3
)
}
\arguments{
\item{MarvelObject}{Marvel object. S3 object generated from \code{CheckAlignment.10x} function.}

\item{coord.intron}{Character string. Coordinates of splice junction whose expression will be plotted.}

\item{x.column}{Character string. Column name in \code{MarvelObject$sample.metadata} to indicate main-cell groups.}

\item{y.column}{Character string. Column name in \code{MarvelObject$sample.metadata} to indicate sub-cell groups.}

\item{x.levels}{Vector of character strings. Main cell groups to include for plotting.}

\item{y.levels}{Vector of character strings. Sub-cell groups to include for plotting.}

\item{xlabels.size}{Numeric value. Font size of heatmap column (main cell group) labels. Default is \code{8}.}

\item{ylabels.size}{Numeric value. Font size of heatmap row (sub-cell group) labels. Default is \code{8}.}

\item{min.pct.cells.gene.expr}{Numeric value. Percentage of cell expressing the gene in a pseudobulk, below which, the value be re-coded as missing and appear as grey on the heatmap.}

\item{min.n.cells}{Numeric value. Minimum number of cells in a cell group, below which, the cell group will be censored and appear as grey on the heatmap.}
}
\value{
An object of class S3 with new slots \code{MarvelObject$adhocPlot$Heatmap$Pseudobulk$Plot} and \code{MarvelObject$adhocPlot$Heatmap$Pseudobulk$Data}.
}
\description{
Heatmap of pseudo-bulk PSI values. x-axis represent main cell group, e.g. donor. y-axis represent sub-cell group, e.g. cell type
}
