% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Script_PLATE_00_rMATS_06_CreateBedFile_RI.R
\name{PrepareBedFile.RI}
\alias{PrepareBedFile.RI}
\title{Preprocess rMATS retained intron (RI) coordinates for Bedtools}
\usage{
PrepareBedFile.RI(file)
}
\arguments{
\item{file}{Data frame. Tab-delimited file output from rMATS. Typically the file is named \code{fromGTF.RI.txt}.}
}
\value{
An object of class data frame that may be used as input for MARVEL
}
\description{
Preprocess rMATS retained intron (RI) coordinates into a BED file suitable for counting intronic coverage with Bedtools
}
