% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Script_PLATE_00_rMATS_03_CreateFeatureData_RI.R
\name{Preprocess_rMATS.RI}
\alias{Preprocess_rMATS.RI}
\title{Preprocess rMATS retained intron (RI) coordinates}
\usage{
Preprocess_rMATS.RI(file, GTF)
}
\arguments{
\item{file}{Data frame. Tab-delimited file output from rMATS. Typically the file is named \code{fromGTF.RI.txt}.}

\item{GTF}{Data frame. The same GTF file used in the rMATS step.}
}
\value{
An object of class data frame that may be used as input for MARVEL
}
\description{
Preprocess rMATS retained intron (RI) coordinates for MARVEL input and annotate the gene type for each gene using the GTF provided.
}
