% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Script_PLATE_00_rMATS_02_CreateFeatureData_MXE.R
\name{Proproces_rMATS.MXE}
\alias{Proproces_rMATS.MXE}
\title{Preprocess rMATS mutually exclusive exons (MXE) coordinates}
\usage{
Proproces_rMATS.MXE(file, GTF)
}
\arguments{
\item{file}{Data frame. Tab-delimited file output from rMATS. Typically the file is named \code{fromGTF.MXE.txt}.}

\item{GTF}{Data frame. The same GTF file used in the rMATS step.}
}
\value{
An object of class data frame that may be used as input for MARVEL
}
\description{
Preprocess rMATS mutually exclusive exons (MXE) coordinates for MARVEL input and annotate the gene type for each gene using the GTF provided.
}
