% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Script_PLATE_00_rMATS_01_CreateFeatureData_SE.R
\name{Proproces_rMATS.SE}
\alias{Proproces_rMATS.SE}
\title{Preprocess rMATS skipped-exon (SE) coordinates}
\usage{
Proproces_rMATS.SE(file, GTF)
}
\arguments{
\item{file}{Data frame. Tab-delimited file output from rMATS. Typically the file is named \code{fromGTF.SE.txt}.}

\item{GTF}{Data frame. The same GTF file used in the rMATS step.}
}
\value{
An object of class data frame that may be used as input for MARVEL
}
\description{
Preprocess rMATS skipped-exon (SE) coordinates for MARVEL input and annotate the gene type for each gene using the GTF provided.
}
