% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Script_PLATE_03_DETECT_EVENTS_3_2_RemoveCrypticSS_AFE.R
\name{RemoveCrypticSS.AFE}
\alias{RemoveCrypticSS.AFE}
\title{Remove cryptic alternative 5' splice site (A5SS)}
\usage{
RemoveCrypticSS.AFE(MarvelObject, DistanceToCanonical = 100)
}
\arguments{
\item{MarvelObject}{Marvel object. S3 object generated from \code{CreateMarvelObject} function.}

\item{DistanceToCanonical}{Numeric value. Distance between alternative and canonical splice site, below which, the alternative splice site will be defined as cryptic and filtered out.}
}
\value{
An object of class S3 with updated slot \code{$SpliceFeature$AFE} and new slot \code{$DistanceToCanonical$AFE}.
}
\description{
Remove cryptic A5SS splicing events from alternative first exon (AFE) splicing evnts whose distance to the canonical are within a user-defined distance.
}
