% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Script_PLATE_03_DETECT_EVENTS_3_1_RemoveCrypticSS_ALE.R
\name{RemoveCrypticSS.ALE}
\alias{RemoveCrypticSS.ALE}
\title{Remove cryptic alternative 3' splice site (A3SS)}
\usage{
RemoveCrypticSS.ALE(MarvelObject, DistanceToCanonical = 100)
}
\arguments{
\item{MarvelObject}{Marvel object. S3 object generated from \code{CreateMarvelObject} function.}

\item{DistanceToCanonical}{Numeric value. Distance between alternative and canonical splice site, below which, the alternative splice site will be defined as cryptic and filtered out.}
}
\value{
An object of class S3 with updated slot \code{$SpliceFeature$ALE} and new slot \code{$DistanceToCanonical$ALE}.
}
\description{
Remove cryptic A3SS splicing events from alternative last exon (ALE) splicing evnts whose distance to the canonical are within a user-defined distance.
}
