% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Script_PLATE_03_DETECT_EVENTS_3_0_RemoveCrypticSS.R
\name{RemoveCrypticSS}
\alias{RemoveCrypticSS}
\title{Remove cryptic alternative 5' and 3' splice site (A5/3SS)}
\usage{
RemoveCrypticSS(MarvelObject, DistanceToCanonical = 100, EventType)
}
\arguments{
\item{MarvelObject}{Marvel object. S3 object generated from \code{CreateMarvelObject} function.}

\item{DistanceToCanonical}{Numeric value. Distance between alternative and canonical splice site, below which, the alternative splice site will be defined as cryptic and removed.}

\item{EventType}{Character string. Type of splicing event. Only applicable to \code{"A5SS"} or \code{"A3SS"}.}
}
\value{
An object of class S3 with updated slot \code{$SpliceFeature$AFE} or \code{$SpliceFeature$ALE} and new slot \code{$DistanceToCanonical$AFE} or \code{$DistanceToCanonical$ALE}.
}
\description{
Subset A5SS and A3SS splicing events from alternative first and last exon (AFE, ALE), respectively, whose distance to the canonical are within a user-defined distance.
}
