% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Script_PLATE_08_PCA_4_RunPCA_PSI_Gene.R
\name{RunPCA.PSI.Exp}
\alias{RunPCA.PSI.Exp}
\title{Integrated splicing and gene non-linear dimension reduction analysis}
\usage{
RunPCA.PSI.Exp(
  MarvelObject,
  n.dim = 20,
  seed = 42,
  sample.ids = NULL,
  cell.group.column,
  cell.group.order,
  cell.group.colors = NULL,
  point.size = 0.5,
  point.alpha = 0.75,
  point.stroke = 0.1
)
}
\arguments{
\item{MarvelObject}{Marvel object. S3 object generated from \code{RunPCA} function.}

\item{seed}{Numeric value. To sure reproducibility of analysis. Default value is \code{42}.}

\item{sample.ids}{Character strings. Specific cells to plot.}

\item{cell.group.column}{Character string. The name of the sample metadata column in which the variables will be used to label the cell groups on the PCA.}

\item{cell.group.order}{Character string. The order of the variables under the sample metadata column specified in \code{cell.group.column} to appear in the PCA cell group legend.}

\item{cell.group.colors}{Character string. Vector of colors for the cell groups specified for PCA analysis using \code{cell.type.columns} and \code{cell.group.order}. If not specified, default \code{ggplot2} colors will be used.}

\item{point.size}{Numeric value. Size of data points on reduced dimension space.}

\item{point.alpha}{Numeric value. Transparency of the data points on reduced dimension space. Take any values between 0 to 1. The smaller the value, the more transparent the data points will be.}

\item{point.stroke}{Numeric value. The thickness of the outline of the data points. The larger the value, the thicker the outline of the data points.}

\item{n.dim.}{Indicate the first number of principal components to use for analysis . Default value is \code{20}, i.e., the first 20 PCs.}
}
\value{
An object of class S3 containing with new slots \code{ MarvelObject$PCA$PSI$ElbowPlot} or \code{MarvelObject$PCA$Exp$ElbowPlot} when \code{level} option set to \code{"splicing"} or \code{"gene"}, respectively.
}
\description{
Non-linear dimension reduction analysis based on both splicing and gene expression data.
}
