% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Script_PLATE_02_SUBSET_CRYPTIC_SS_2_A3SS.R
\name{SubsetCrypticSS.A3SS}
\alias{SubsetCrypticSS.A3SS}
\title{Subset cryptic alternative 3' splice site (A3SS)}
\usage{
SubsetCrypticSS.A3SS(MarvelObject, DistanceToCanonical = 100)
}
\arguments{
\item{MarvelObject}{Marvel object. S3 object generated from \code{CreateMarvelObject} function.}

\item{DistanceToCanonical}{Numeric value. Distance between alternative and canonical splice site, below which, the alternative splice site will be defined as cryptic and retained.}
}
\value{
An object of class S3 with updated slot \code{$SpliceFeature$A3SS} and new slot \code{$DistanceToCanonical$A3SS}.
}
\description{
Subset A3SS splicing events whose distance to the canonical are within a user-defined distance.
}
