% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Script_PLATE_02_SUBSET_CRYPTIC_SS_1_A5SS.R
\name{SubsetCrypticSS.A5SS}
\alias{SubsetCrypticSS.A5SS}
\title{Subset cryptic alternative 5' splice site (A5SS)}
\usage{
SubsetCrypticSS.A5SS(MarvelObject, DistanceToCanonical = 100)
}
\arguments{
\item{MarvelObject}{Marvel object. S3 object generated from \code{CreateMarvelObject} function.}

\item{DistanceToCanonical}{Numeric value. Distance between alternative and canonical splice site, below which, the alternative splice site will be defined as cryptic and retained.}
}
\value{
An object of class S3 with updated slot \code{$SpliceFeature$A5SS} and new slot \code{$DistanceToCanonical$A5SS}.
}
\description{
Subset A5SS splicing events whose distance to the canonical are within a user-defined distance.
}
