% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Script_PLATE_02_SUBSET_CRYPTIC_SS_0.R
\name{SubsetCrypticSS}
\alias{SubsetCrypticSS}
\title{Subset cryptic alternative 5' and 3' splice site (A5/3SS)}
\usage{
SubsetCrypticSS(MarvelObject, DistanceToCanonical = 100, EventType)
}
\arguments{
\item{MarvelObject}{Marvel object. S3 object generated from \code{CreateMarvelObject} function.}

\item{DistanceToCanonical}{Numeric value. Distance between alternative and canonical splice site, below which, the alternative splice site will be defined as cryptic and retained.}

\item{EventType}{Character string. Type of splicing event. Only applicable to \code{"A5SS"} or \code{"A3SS"}.}
}
\value{
An object of class S3 with updated slot \code{$SpliceFeature$A5SS} or \code{$SpliceFeature$A3SS} and new slot \code{$DistanceToCanonical$A5SS} or \code{$DistanceToCanonical$A3SS}.
}
\description{
Subset A5SS and A3SS splicing events whose distance to the canonical are within a user-defined distance.
}
