% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Script_DROPLET_02_PREPROCESS_3_ValidateSJ.R
\name{ValidateSJ.10x}
\alias{ValidateSJ.10x}
\title{Validate splice junctions}
\usage{
ValidateSJ.10x(MarvelObject)
}
\arguments{
\item{MarvelObject}{Marvel object. S3 object generated from \code{AnnotateSJ.10x} function.}
}
\value{
An object of class S3 containing the updated slots \code{MarvelObject$sj.metadata} and \code{MarvelObject$sj.count.matrix}.
}
\description{
Retains splice junctions whose start and end belong to the same gene.
}
