% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Script_PLATE_04_PREPROCESS_3_3_CheckAlignment_PSI_Gene.R
\name{CheckAlignment.PSI.Exp}
\alias{CheckAlignment.PSI.Exp}
\title{Check splicing and gene data against each other}
\usage{
CheckAlignment.PSI.Exp(MarvelObject)
}
\arguments{
\item{MarvelObject}{S3 object generated from \code{CheckAlignment.PSI} and \code{CheckAlignment.Exp} function.}
}
\value{
An object of class S3 with updated slots \code{MarvelObject$SplicePheno}, \code{MarvelObject$PSI}, \code{MarvelObject$GenePheno}, and \code{MarvelObject$Exp}.
}
\description{
Subsets overlapping samples between splicing and gene data.
}
\examples{
marvel.demo <- readRDS(system.file("extdata/data", "marvel.demo.rds", package="MARVEL"))

marvel.demo <- CheckAlignment.PSI.Exp(MarvelObject=marvel.demo)
}
