% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Script_PLATE_02_COMPUTE_PSI_3_Compute_A5SS.R
\name{ComputePSI.A5SS}
\alias{ComputePSI.A5SS}
\title{Compute Alternative 5' Splice Site (A5SS) Percent Spliced-in (PSI) Values}
\usage{
ComputePSI.A5SS(MarvelObject, CoverageThreshold)
}
\arguments{
\item{MarvelObject}{S3 object generated from \code{CreateMarvelObject} function.}

\item{CoverageThreshold}{Numeric value. Coverage threshold below which the PSI of the splicing event will be censored, i.e. annotated as missing (NA). Coverage defined as the total number of reads supporting both included and excluded isoforms.}
}
\value{
An object of class S3 containing all the original slots as inputted by the user in addition to two new slots. \code{$SpliceFeatureValidated$A5SS} contains the validated splicing event metadata. \code{$PSI$A5SS} contains the computed PSI values for the validated splicing events.
}
\description{
\code{ComputePSI.A5SS} computes percent spliced-in (PSI) alternative 5' splice site (A5SS) splicing event.
}
\details{
This function computes the PSI for each A5SS splicing event. Splicing events provided in \code{SpliceFeature} data frame will first be cross-checked against the splice junctions provided in \code{SpliceJunction} data frame. Only events whose junctions are found in \code{SpliceJunction} are retained. The formula for computing PSI is the number of junction reads supporting the included isoform divided by the total number of reads supporting both included and excluded isoforms.
}
\examples{
# Load input
marvel <- readRDS(system.file("extdata/Data", "MarvelObject.rds", package="MARVEL"))

# Run example
marvel <- ComputePSI.A5SS(MarvelObject=marvel,
                          CoverageThreshold=10
                          )

# Check output
marvel$SpliceFeatureValidated$A5SS
marvel$PSI$A5SS
}
\author{
Sean Wen <sean.wenwx@gmail.com>
}
