% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Script_DROPLET_07_ADHOC_PLOT_BOXPLOT_1_PlotValues_PSI_PseudoBulk.R
\name{PlotValues.PSI.Pseudobulk.10x}
\alias{PlotValues.PSI.Pseudobulk.10x}
\title{Boxplot of pseudo-bulk PSI values vs cell groups}
\usage{
PlotValues.PSI.Pseudobulk.10x(
  MarvelObject,
  cell.group.list,
  coord.intron,
  cell.group.colors = NULL,
  xlabels.size = 10,
  min.pct.cells.gene.expr = 10
)
}
\arguments{
\item{MarvelObject}{Marvel object. S3 object generated from \code{CheckAlignment.10x} function.}

\item{cell.group.list}{List of character strings. Each element in the list (main list) is a list (sub-list). Each sub-list in turn is a list of pseudobulk samples represented by cell IDs. Each element in the main list is a cell group and the label of each element of the main list represents the cell group label.}

\item{coord.intron}{Character string. Coordinates of splice junction whose expression will be plotted.}

\item{cell.group.colors}{Vector of character strings. Colors of cell groups and should be same length as \code{cell.group.list}. Default \code{ggplot2} colors are used.}

\item{xlabels.size}{Numeric value. Font size of x-tick labels. Default is \code{10}.}

\item{min.pct.cells.gene.expr}{Numeric value. Percentage of cell expressing the gene in a pseudobulk, below which, the pseudobulk will be omitted from plotting.}
}
\value{
An object of class S3 with new slots \code{MarvelObject$adhocPlot$Boxplot$Pseudobulk$Plot} and \code{MarvelObject$adhocPlot$Boxplot$Pseudobulk$Data}.
}
\description{
Boxplot of pseudo-bulk PSI values on the y-axis against cell groups on the x-axis.
}
